fix: #33 Remove unsupported message and refactor display of URL Signed-off-by: Laurent Broudoux <laurent.broudoux@gmail.com>
